durable-streaming/pyproject.toml
[project]
name = "durable-streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.14",
    "uvicorn>=0.35.0",
    "sqlmodel>=0.0.22",
    "mirascope[openai]>=1.25.2",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
